requirements.txt
Django==1.7.1
pycrypto==2.6.1
requests==2.6.0
wsgiref==0.1.2
